backpropagate condition on t
the condition that a family of loops must be constant for t <= 0 or t >= 1 was required for nice_loop, but not for surrounded_family.
To keep it this way meant that I had to truncate t multiple times in the proof of exists_loops.
To simplify and uniformize, I also added this condition (or more precisely: an equivalent one) to surrounding_family