redefine one_jet_sec
The new definition is more convenient, since an arbitrary one_jet_sec is now a section up to refl.
You also have to specify slightly less information when defining one.
This is one way to have to deal less with casts, and I'm wondering if you should use this more throughout mathlib