Move `diffeomorph_to_nhd` to its own file.
Previously this provided a `smooth_open_embedding` and so could not be
so easily decoupled. However with the recent refactor to avoid using
`model_with_corners_self` it became more convenient for `diffeomorph_to_nhd`
to be stated in terms of `cont_diff` instead of `cont_mdiff` and thus
it stopped being a `smooth_open_embedding`.